% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2nmode.R
\name{df2nmode}
\alias{df2nmode}
\title{df2nmode}
\usage{
df2nmode(df, n = 4)
}
\arguments{
\item{df}{the dataframe}

\item{n}{the number of columns. Last column equals edgeweight}
}
\value{
returns edgelist
}
\description{
transforms (wide) `dataframe` into (long) edgelists to work with `igraph`
}
\examples{
df <- data.frame(from = 1:4,
                 to = c("a", "b", "b", "a"),
                 frame = paste0("f", 4:1),
                 time = paste0("t", c(1,1,2,2)),
                 weight = 1,
                 stringsAsFactors = FALSE)

# 1st example
df2nmode(df, n = 4)

# 2nd example
for (i in 3:5) {
          x <- df2nmode(df, n = i)
          g <- igraph::graph_from_data_frame(x, directed = FALSE)
          plot(g, main = paste("ncol equal to", i))
          Sys.sleep(2)
}
}
