% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(source = "online")
}
\arguments{
\item{source}{either "online" (default) to download data or "example" to access inbuild data set}
}
\value{
(optionally) returns selected data set as .csv or `dataframe`
}
\description{
starts shiny app with GUI for data selection and visualization
}
\examples{
\dontrun{
get_data(source = "example")
}
}
