% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{get_predictions}
\usage{
get_predictions(text, output = "word", verbose = F, sleep = 0)
}
\arguments{
\item{text}{character input of length 1}

\item{output}{set to "word" (default) for BIO output or "sentence" for sentence level output}

\item{verbose}{logical; print flow of information (default FALSE)}

\item{sleep}{rest time between requests (in seconds; defaults to zero)}
}
\value{
returns dataframe
}
\description{
send text input to API and receive predictions
}
\examples{
\dontrun{
text <- "Seehofer fordert eine Obergrenze. Merkel lehnt eine Obergrenze ab."
df <- get_predictions(text = text, output = "sentence")
}
}
