% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_codes.R
\name{lookup_codes}
\alias{lookup_codes}
\title{lookup_codes}
\usage{
lookup_codes(code, language = "english")
}
\arguments{
\item{code}{the code category as string or integer}

\item{language}{set to return 'english' (default) or 'german' category names}
}
\value{
returns vector containing category names. Multiple queries are possible.
}
\description{
convert codes into category names
}
\examples{
 lookup_codes(code = 102)
 lookup_codes(c(303, 101, "801"), language = "german")
}
